switch to qt-wrapper-as-a-file
See #7128 for discussion. Mypy doesn't see to follow wildcard imports
like those used in /qt/gui.py etc. Importing the Qt modules as
attributes seems to work better for that. It does mean we can't import
stuff from Qt modules though. We have to import the module and then
reference the things in it as attributes.

The _Machinery class and renaming at the bottom of the file are quick
hacks to make it compatible with the qts-like approach. Mostly this
qt.py file is copied from a PyQt5-only version and it could use a little
refactoring in it's current state.